dashboard - increase the distance between title and subtitle on summary widget Fixes #72
