+++ lib/ui/common/dashboard.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:howmuch/ui/dashboard/account_list.dart';
+import 'package:howmuch/ui/dashboard/summary.dart';
 import 'package:intl/intl.dart';
 import 'package:line_awesome_icons/line_awesome_icons.dart';
 import 'package:liquid_progress_indicator/liquid_progress_indicator.dart';
@@ -17,8 +18,6 @@ import 'custom_card.dart'; // https://pub.dev/packages/fluid_layout
 
 final _numberFormat =
     NumberFormat.currency(locale: 'de_CH', symbol: '', decimalDigits: 2);
-final _percentFormat =
-    NumberFormat.decimalPercentPattern(locale: 'de_CH', decimalDigits: 2);
 
 class Dashboard extends StatefulWidget {
   @override
@@ -256,131 +255,6 @@ class LeftLayoutWidget extends StatelessWidget {
   }
 }
 
-class SummaryWidget extends StatelessWidget {
-  final String widgetType;
-  final double amount, variation;
-  SummaryWidget(
-      {@required this.widgetType,
-      @required this.amount,
-      @required this.variation});
-
-  @override
-  Widget build(BuildContext context) {
-    String subtitle, variationTitle, periodLabel;
-    IconData icon, iconVariation;
-    Color colorVariation;
-
-    // period selected in the top widget of the dashboard
-    var _periodValue = Provider.of<DashboardFilter>(context).getPeriodFilter;
-    switch (_periodValue) {
-      case 1:
-        periodLabel = 'Week';
-        break;
-      case 2:
-        periodLabel = 'Month';
-        break;
-      case 3:
-        periodLabel = 'Year';
-        break;
-      default:
-        periodLabel = 'N/A';
-    }
-
-    colorVariation =
-        this.variation < 0 ? Color(debitColor) : Color(creditColor);
-    iconVariation = this.variation < 0
-        ? LineAwesomeIcons.chevron_circle_down
-        : LineAwesomeIcons.chevron_circle_up;
-
-    if (this.widgetType == 'INCOME') {
-      subtitle = 'Income In this ' + periodLabel;
-      variationTitle = 'Balance Up by';
-      icon = LineAwesomeIcons.battery_full;
-    }
-    if (this.widgetType == 'OUTCOME') {
-      subtitle = 'Outcome In this ' + periodLabel;
-      variationTitle = 'Balance Down by';
-      icon = LineAwesomeIcons.battery_quarter;
-    }
-    if (this.widgetType == 'SAVINGS') {
-      subtitle = 'Saved In this ' + periodLabel;
-      variationTitle = 'Saved Up by';
-      icon = LineAwesomeIcons.leaf;
-    }
-
-    return Padding(
-      padding: const EdgeInsets.fromLTRB(16.0, 16.0, 16.0, 6.0),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Padding(
-                padding: const EdgeInsets.only(top: 10.0),
-                child: Icon(icon, size: 36),
-              ),
-              Padding(
-                padding: const EdgeInsets.only(left: 20.0),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.start,
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Text(
-                      'CHF  ' + _numberFormat.format(this.amount),
-                      style: TextStyle(
-                        fontSize: 24,
-                        fontWeight: FontWeight.bold,
-                      ),
-                    ),
-                    Text(
-                      subtitle,
-                      style: TextStyle(
-                        fontSize: 14,
-                        color: Color(buttonColor),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.start,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Text(variationTitle,
-                  style: TextStyle(
-                    fontSize: 14,
-                    color: Color(buttonColor),
-                  )),
-              Spacer(),
-              Padding(
-                padding: const EdgeInsets.only(right: 6.0),
-                child: Icon(
-                  iconVariation,
-                  size: 22,
-                  color: colorVariation,
-                ),
-              ),
-              Text(
-                _percentFormat.format(this.variation),
-                style: TextStyle(
-                  color: colorVariation,
-                  fontWeight: FontWeight.bold,
-                ),
-              )
-            ],
-          )
-        ],
-      ),
-    );
-  }
-}
-
 class HeaderLayoutWidget extends StatelessWidget {
   HeaderLayoutWidget();
 
@@ -400,7 +274,7 @@ class HeaderLayoutWidget extends StatelessWidget {
               child: Padding(
                 padding: const EdgeInsets.all(16.0),
                 child: CustomCard(
-                  child: SummaryWidget(
+                  child: Summary(
                     widgetType: 'INCOME',
                     amount: FakeData.fakeIncome,
                     variation: FakeData.fakeIncomeVariation / 100.0,
@@ -416,7 +290,7 @@ class HeaderLayoutWidget extends StatelessWidget {
               child: Padding(
                 padding: const EdgeInsets.all(16.0),
                 child: CustomCard(
-                  child: SummaryWidget(
+                  child: Summary(
                     widgetType: 'OUTCOME',
                     amount: FakeData.fakeOutcome,
                     variation: FakeData.fakeOutcomeVariation / 100.0,
@@ -432,7 +306,7 @@ class HeaderLayoutWidget extends StatelessWidget {
               child: Padding(
                 padding: const EdgeInsets.all(16.0),
                 child: CustomCard(
-                  child: SummaryWidget(
+                  child: Summary(
                     widgetType: 'SAVINGS',
                     amount: FakeData.fakeSavings,
                     variation: FakeData.fakeSavingsVariation / 100.0,
